- string class, interface only for now
Well I think I'll base this one on std::vector<charT>. Given proper inlining by the compiler it shouldn't come with any performance penalty. If performance is lacking it means I need a vector specialization for PODish types.

I think I will implement small string optimization, because it just makes sense. The minimal size of string/vector is 24 bytes, I probably need one pointer to be 0 to signal that SSO is in use, but this leaves 15/16 bytes available to store characters without any heap allocation and/or growing the handle.